Refactor Componentable methods into Recreatable
