+++ pitop/core/mixins/recreatable.py
@@ -1,4 +1,5 @@
 import importlib
+import json
 
 
 class Recreatable:
@@ -8,19 +9,6 @@ class Recreatable:
         self.add_to_config("module", self.__module__)
         for k, v in config_dict.items():
             self.add_to_config(k, v)
-
-    def is_builtin_class_instance(self, obj):
-        return obj.__class__.__module__ == '__builtin__'
-
-    def add_to_config(self, key, value):
-        self._config[key] = value
-
-    @property
-    def component_config(self):
-        cfg = {}
-        for k, v in self._config.items():
-            cfg[k] = v() if callable(v) else v
-        return cfg
 
     @classmethod
     def from_dict(cls, config_dict):
@@ -34,7 +22,32 @@ class Recreatable:
         main_cls = cls.import_class(module_name, cls_name)
         return main_cls(**config_dict)
 
+    @classmethod
+    def from_file(cls, path):
+        print(f"Loading configuration from {path}")
+        with open(path) as json_file:
+            config_dict = json.load(json_file)
+        return cls.from_dict(config_dict)
+
+    def save_config(self, path):
+        print(f"Storing configuration in {path}")
+        with open(path, "w") as writer:
+            json.dump(self.component_config, writer, indent=4)
+
+    def add_to_config(self, key, value):
+        self._config[key] = value
+
+    @property
+    def component_config(self):
+        cfg = {}
+        for k, v in self._config.items():
+            cfg[k] = v() if callable(v) else v
+        return cfg
+
     @staticmethod
     def import_class(module_name, class_name):
         module = importlib.import_module(module_name)
         return getattr(module, class_name)
+
+    def print_config(self):
+        print(json.dumps(self.component_config, indent=4))
